Add setup.cfg with sensible flake8 configuration
This is a sensible config which allows to run "flake8 ." without having
too many useless warnings.

Note that the `per-file-ignores` config requires flake8 >= 3.7.

Ignore:

- E501, line too long, because we don't really care about that (black
  decides how it's formatted anyway).
- W503, line break before bbinary operator, because we don't really care
  about that (black decides how it's formatted anyway).

__init__.py contains some expectedly unused imports, so it seems cleaner
to disable that warning for the whole file, rather than putting a "noqa"
comment on each line.  Of course, we risk having a really unused import
in that file, but it doesn't contain much, so the risk is low.

Finally, exclude the Python tap directory, as it's not our code.

Change-Id: Iebdf274b91907fb31d4ebfa6ae4c8157de46467c
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/2083
Reviewed-by: Francis Deslauriers <francis.deslauriers@efficios.com>