bt2: remove duplicate test test_create_from_vint in RealValueTestCase
There are two tests named like this in RealValueTestCase.  The second
one looks right (it creates a Value from an int, then creates a
RealValue from that).  The first one looks like it should be called
"test_create_from_vfloat", as it tests creating a RealValue from another
RealValue, but we already have such a test case a bit lower.  So I think
it's fine to delete it.

Reported-by: flake8
Change-Id: I274b922c653675913f2346cbc67cbcd03cd5208f
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/2086
Reviewed-by: Francis Deslauriers <francis.deslauriers@efficios.com>