tests: convert paths passed to lttng_live_server.py
On Windows, the MSYS2 bash tries to convert paths in the Unix form
(/c/foo/bar) to the Windows form (C:\foo\bar).  However, some paths are
embedded in some complex arguments we pass to lttng_live_server.py, and
bash does not convert them.  This causes the file open in Python to
fail, because it doesn't find that file.  We therefore need to convert
them by hand using cygpath before passing them to the Python script.

Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Change-Id: I108e6dc2a690042bf727335fc5af5294bf928590
Reviewed-on: https://review.lttng.org/c/babeltrace/+/2112
Reviewed-by: Michael Jeanson <mjeanson@efficios.com>
Tested-by: jenkins <jenkins@lttng.org>