cli: update error message in set_stream_intersections
The error message printed when the `babeltrace.trace-infos` query fails,
regardless of the reason, is "Component class does not support the
`babeltrace.trace-infos` query".  I think it's misleading, because this
is just one particular error case.  It could be that the component class
supports the query, but failed because it didn't like the parameters we
passed.

Update the message to be a bit more generic.  The fail_reason printed
just after gives a bit more precision about the error.

Change-Id: If556129b612cb6aa9c480b3f863368e48e8a6d4a
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/2115
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>