tests: add format attributes to functions receiving format strings in tap.h
This makes the compiler validate the format strings against the types of
the passed arguments.  It found a few (minor) issues, which are also
fixed by this patch.

Note that in test_bitfield.c, the format string now includes "0x" before
printing the number in hexadecimal, to avoid any confusion when the user
reads it.

Change-Id: I07cac88aa3cdd445d79f2c12bc0f9333f6a768a9
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/2129
Tested-by: jenkins <jenkins@lttng.org>