flt.lttng-utils.debug-info: add all SC and EC to output trace class ASAP
The fact that we are copying metadata only when encountering the data
associated to it leads to variations in the output of a
`sink.text.details` component when changing message batch size. It is
not wrong it's simply annoying.

So to improve that, this commit makes `flt.lttng-utils.debug-info`
components copy all the metadata of the input trace class as soon as it
encounters it.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>
Change-Id: I7a58e3ed4659ee73f5b0f94da5427ac8b06d5178
Reviewed-on: https://review.lttng.org/c/babeltrace/+/2132
Tested-by: jenkins <jenkins@lttng.org>