Cleanup: flt.lttng-utils.debug-info: coding style
When I first wrote this component class, I was using the wrong coding
style standard. More specifically, the main problem is that the number
of tabs on breaking lines is wrong at bunch places in the debug-info
files.

I am currently working on this code and it annoyed me. So I fixed most
of it.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>
Change-Id: I41b3d6b3a35546e5ce4f65470bed1303966523c4
Reviewed-on: https://review.lttng.org/c/babeltrace/+/2134
Tested-by: jenkins <jenkins@lttng.org>
Reviewed-by: Simon Marchi <simon.marchi@efficios.com>