flt.lttng-utils.debug-info: tidy up component initialization
Assign add_port_status to status directly, as has been the accepted
pattern in the rest of the project for a while.  This is less verbose
than using a switch-case.

Make init_from_params return a
bt_component_class_initialize_method_status, since it's only used in the
context of component initialization.  The alert reader will notice that
this function can't fail, making it pointless to have it return a
status.  However, it will be able to fail in the next patch, so I chose
to leave a non-void return type.

Change-Id: I41c1f6dc9dea5cbda4ef33f336d4b2cff216a568
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/2141
Tested-by: jenkins <jenkins@lttng.org>