src.text.dmesg: validate parameters using param-validation
This patch makes src.text.dmesg validate its parameters using the
param-validation utility, instead of doing it by hand.

The

   Cannot specify both `read-from-stdin` and `path` parameters.

error seems unreachable, as there's no way that read_from_stdin can be
true when we read the path parameter.  I presume that there used to be a
read-from-stdin parameter, making this situation possible, but that it
was eventually removed.

Change-Id: Ie4c82968ba0aae24fc479a0f4442c92d50c489e4
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/2147