tests: test_field.py: enable tests using _test_binop_lhs_value_same
Tests using _test_binop_lhs_value_same, that make sure that using a
binary operator on a field doesn't change that field's value, are
currently disabled.  The problem is this line, that tries to copy the
field:

    value_before = copy.copy(self._def)

Since we have removed everything related to copy of fields, this is
unsupported.  However, the same test but for unary operators copies the
Python value instead.  Change _test_binop_lhs_value_same to do the same.

It's not clear to me if copying the value is really necessary, but I
prefer to leave it like this at the moment.

Change-Id: Ib35cbb44c000293e7e8ba42c020b5ace649a268d
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/2205
Tested-by: jenkins <jenkins@lttng.org>
Reviewed-by: Francis Deslauriers <francis.deslauriers@efficios.com>