Bump required glib version to 2.28
While investigating a hang in the lttng-live source component when
running with glib 2.22.5, we have found that this version had a bug in
g_ptr_array_remove_index_fast.  When removing the last element of the
array, the free func is not called.

It was fixed in the following commit:

https://github.com/GNOME/glib/commit/5fffa39b6ae8f8faf1036fbf07de02ffe84ef099

The first major release where this is fixed is 2.26.  We went through
the platforms we need to support, and the oldest glib version we need to
support is 2.28.6, on Solaris 11.  Therefore, fix the issue with
lttng-live by bumping the required glib version to 2.28.

Change-Id: Ib56ac221dfc5ecf45eb62b837fe0b6f54032cf31
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/2210
Reviewed-by: Michael Jeanson <mjeanson@efficios.com>
Reviewed-by: Jonathan Rajotte Julien <jonathan.rajotte-julien@efficios.com>
Tested-by: jenkins <jenkins@lttng.org>